src/test/java/com/baseMVC/services/UserServiceTest.java
package com.baseMVC.services;

import com.baseMVC.config.AppConfig;
import com.baseMVC.entities.User;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.web.WebAppConfiguration;

import java.util.List;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes = AppConfig.class)
@WebAppConfiguration
public class UserServiceTest {

    @Autowired
    private UserService userService;

    @Test
    public void findAllUsers()  {
        List<User> users = userService.findAll();
        assertNotNull(users);
        assertTrue(!users.isEmpty());
    }

    @Test
    public void testFindByEmail() throws Exception {
        User user = userService.findByEmail("foratala@gmail.com");
        assertNotNull(user);
    }
}
